kernel: carry the full 64-bit signal mask through the signal frame (sigreturn)
The deliver->sigreturn round-trip stored the saved signal mask in a 32-bit
SigContext field and reassembled it through 32-bit arch contracts, truncating
the 64-bit blocked mask. Any handler invocation silently lost the block state
for signals 33..64 (including SIGCANCEL=32, bit 31). Widen SigContext with an
oldmask_hi word and make archPushSignalFrame/archSigreturn carry uint64_t end
to end, so the full mask survives across the handler.

+++ arch/include/arch/usermode.h
@@ -50,11 +50,11 @@ enum {
 // sigreturn. `restartAction` (SIG_FRAME_*) decides what the resumed context does about an
 // interrupted syscall; `origEax` is the syscall number to restore when restarting.
 void archPushSignalFrame(TrapFrame* tf, uint32_t handler, uint32_t restorer,
-                         int sig, uint32_t oldMask, uint32_t origEax, int restartAction);
+                         int sig, uint64_t oldMask, uint32_t origEax, int restartAction);
 
-// SYS_sigreturn: restore `tf` from the user-stack signal frame; writes the mask to be
-// restored to *oldMaskOut and returns the interrupted code's saved eax.
-int archSigreturn(TrapFrame* tf, uint32_t* oldMaskOut);
+// SYS_sigreturn: restore `tf` from the user-stack signal frame; writes the full 64-bit
+// mask to be restored to *oldMaskOut and returns the interrupted code's saved eax.
+int archSigreturn(TrapFrame* tf, uint64_t* oldMaskOut);
 
 // The current syscall result sitting in the trap frame (eax), as a signed int.
 int archSyscallResult(TrapFrame* tf);
